Updated module handling and added network checks
In `builder.py`, the `sys` module was moved from `option_module_mapping_pyinstaller` to `included_modules_pyinstaller`. In `luna.py`, the `sys` module was imported and a new function `IsConnectedToInternet` was added to the `Luna` function to check internet connectivity. If no connection is found, the program will exit. In `Debug.py`, the `random` module and `PoolManager` from `urllib3` were imported. A new method `checkHTTPSimulation` was added to the `Debug` class to simulate an HTTP request and return a boolean based on the success of the request.