`Added GetSelf method and updated delete method in SelfDestruct class`
Implemented a new `GetSelf` method in the `SelfDestruct` class to determine if the script is running as a compiled executable or as a python script. The `delete` method was updated to use `GetSelf` for determining the script's path and type. If the script is a compiled executable, it is deleted using a hidden subprocess. If not, it is removed using `os.remove`. The old code for creating and running a batch file to delete the script and handling exceptions was removed.