`Added EntryPoint renaming for enhanced security`
` `

`Added a new method `RenameEntryPoint` to rename the entry point of the executable file. This method is called after the `RemoveMetaData` method in both the `pyinstaller` and `nuitka` sections of the code, with the new entry point name being "Luna". These changes aim to enhance the security of the compiled executable file by randomizing its entry point, making it more difficult for malicious actors to predict and exploit.`