Refactor method order and rename variables in builder.py and Debug.py
In builder.py, the order of `RenameEntryPoint` and `AddCertificate` methods has been swapped for better flow. In Debug.py, `blacklisted_hwids` attribute has been renamed to `blacklisted_uuids` and a new UUID '00000000-0000-0000-0000-000000000000' has been added to the set. Also, the variable `hwid` in the `get_system` method has been renamed to `uuid` for consistency and clarity.