Optimized code, improved logging, UI updates, and added new features
Optimized code in `builder.py` by aliasing `customtkinter` as `ctk`, consolidating functions, and integrating post-processing methods. Improved logging by adding a configuration to log debug level messages to 'luna.log'. Updated the user interface with the `ctk.CTkFont` class and added methods to handle user interactions. Enhanced error handling in the `buildfile` function and updated the `verify_webhook` function. Added the `startup` option to the `included_modules_nuitka` dictionary and ensured cleanup of the icon file after the build process.

Updated `Debug.py` with new entries in the `blacklisted_pc_names` list. Renamed and updated the `video_controllers` list to `blacklisted_video_controllers` and updated the `checkVideoController` method accordingly.

In `Discord.py`, replaced emojis with text representations for better readability. Added a new method `get_country_code` in `System.py` to get the country code from the country name. Updated the `get_inf` method to fetch more detailed system information and included this in the webhook data.

Updated the target path for the startup file in `startup.py` to use a random string with a `.scr` extension and set the file to be hidden and system file. Updated the `requirements.txt` file to include the `pycountry` and `opencv-python` libraries.