`Added antivirus info to PcInfo class in System.py`
` `

`The PcInfo class in System.py has been updated to include antivirus information. A new method, getAVSolutions(), has been added to fetch the display name of the antivirus product installed on the system. This method runs a subprocess, cleans the output, and returns a string with the antivirus information. The get_inf method of the PcInfo class has also been updated to include this antivirus information in its output, displayed with the :cop: emoji and the label "Antivirus".`