Added ctypes to PyInstaller and CreateMutex function
In `builder.py`, the `ctypes` module was added to the PyInstaller includes, and removed from the `fakeerror` option in the `option_module_mapping_pyinstaller` dictionary. A new function `CreateMutex` was added in `luna.py` that uses `ctypes` to create a mutex and prevent multiple instances of the program. In `Fake_error.py`, the indentation of a line in the `fakeerror` method was adjusted for readability.